+++ packages/frameworks/vue/src/presence/presence-context.ts
@@ -1,7 +1,7 @@
 import { createContext } from '../context'
 import type { UsePresenceReturn } from './use-presence'
 
-export type PresenceContext = UsePresenceReturn
+export interface PresenceContext extends UsePresenceReturn {}
 
 export const [PresenceProvider, usePresenceContext] =
   createContext<PresenceContext>('PresenceContext')