+++ packages/frameworks/vue/src/presence/presence-props-context.ts
@@ -1,7 +1,7 @@
 import { createContext } from '../context'
 import type { UsePresenceProps } from './use-presence'
 
-export type PresencePropsContext = UsePresenceProps
+export interface PresencePropsContext extends UsePresenceProps {}
 
 export const [PresencePropsProvider, usePresencePropsContext] =
   createContext<PresencePropsContext>('PresencePropsContext')